// cypress/integration/ascend/interfaces/ReadinessAssessmentDashboardInterface.ts
export enum ReadinessAssessmentDashboardBlocks {
    Availability = 'Availability',
    Security = 'Security',
    Confidentiality = 'Confidentiality',
    Privacy = 'Privacy'
}

export type ReadinessAssessmentDashboardInterface = ReadinessAssessmentDashboardBlocks;

export enum ReadinessAssessmenLabels {
    Manage = 'Manage',
    Select = 'Select',
    Continue = 'Continue',
    RequestAssessmentUpgrade = 'Request Assessment Upgrade'
} 